Fixed linkage error caused by callback functions
